Major progress on the awards page
Specifically, the rendered page looks pretty much how I wanted it to
look from the beginning, with clicky-buttons instead of checkboxes.

Furthermore, significant testing work was done for the awards view.
However, my approach was incorrect: currently it is done by mocking the
model layer, which is a great idea in general, but it's going to take
way too much time for a quick-turnaround website with only so many
required test cases. Performance will suffer when I switch to using the
Django test client, but then again, if I wanted performance, I'd write
the whole thing in Go or something.