Added FinancialAid page
The entire financial aid page is now functional. Generating the table
turned out to be relatively complex, and so it is done entirely in the
view. This makes it easier to handle the logic therein (grabbing errors
from elsewhere in the class, dealing with variable-sized cells,
alternating styles, etc.).

The following minor changes were required to make this all happen:
* Wrote code to generate the table, and styled the table
* Added basic javascript (for save button and minus buttons)
* Custom validation for finaid makes better use of the issue field
* New migration to use a semester end_date instead of a real date
* Added a previously missing test for aforementioned end semester