Introduced a (short) DPan scholarship essay
Cameron was right: depending entirely on peer feedback for the Dan
Summers scholarship is asking for trouble, due to the lack of structured
feedback *plus* the lack of basic information from the actual candidate.

However, I have no intention of making the essay for this new award as
in-depth as those for the Ean Hong and the Ambassador, just because it
isn't well-established enough yet.